requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.101.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
pydantic==2.1.1
pydantic_core==2.4.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
